refactor: add additional exception types for passing requestId back to client app, move exceptions into client lib
