+++ libraries/client/src/main/java/org/holochain/androidserviceruntime/client/AdminBinderUnauthorizedException.kt
@@ -1,0 +1,6 @@
+package org.holochain.androidserviceruntime.client
+
+class AdminBinderUnauthorizedException(
+    message: String = "Unauthorized",
+    cause: Throwable? = null,
+) : Exception(message, cause)
