+++ libraries/client/src/main/java/org/holochain/androidserviceruntime/client/AppBinderRequestAuthorizationException.kt
@@ -1,0 +1,6 @@
+package org.holochain.androidserviceruntime.client
+
+class AppBinderRequestAuthorizationException(
+    requestId: String,
+    message: String = "Unauthorized, requesting authorization",
+) : Exception(message, null)
