+++ libraries/client/src/main/java/org/holochain/androidserviceruntime/client/AppBinderUnauthorizedException.kt
@@ -1,6 +1,6 @@
-package org.holochain.androidserviceruntime.service
+package org.holochain.androidserviceruntime.client
 
-class UnauthorizedException(
+class AppBinderUnauthorizedException(
     message: String = "Unauthorized",
     cause: Throwable? = null,
 ) : Exception(message, cause)
