--- libraries/service/src/main/java/org/holochain/androidserviceruntime/service/UniqueNotificationId.kt
@@ -1,9 +1,0 @@
-package org.holochain.androidserviceruntime.service
-
-import java.util.concurrent.atomic.AtomicInteger
-
-public class UniqueNotificationId {
-    private var counter = AtomicInteger()
-
-    public fun get(): Int = counter.getAndIncrement()
-}
